requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 音频处理
pyaudio==0.2.11
soundfile==0.12.1
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4

# 语音识别 (ASR)
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# 文本转语音 (TTS)
TTS==0.20.6
espeak==1.0.3

# 大语言模型
openai==1.3.7
anthropic==0.8.1
httpx==0.25.2

# Live2D和图形
opencv-python==4.8.1.78
Pillow==10.1.0
pygame==2.5.2

# UI界面
tkinter-page==0.0.4
customtkinter==5.2.2

# 配置和日志
PyYAML==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# 异步和并发
asyncio-mqtt==0.13.0
websockets==12.0
aiofiles==23.2.1

# 工具库
typing-extensions==4.8.0
dataclasses-json==0.6.2
marshmallow==3.20.1
requests==2.31.0 